+++ R/tidyterra-package.R
@@ -1,12 +1,15 @@
 #' @keywords internal
 "_PACKAGE"
 
+## usethis namespace: start
+#' @importFrom lifecycle deprecated
 #' @importFrom ggplot2 aes after_stat
 #' @importFrom rlang .data
 #' @importFrom grDevices terrain.colors rgb col2rgb grey
 #' @importFrom tibble as_tibble
 #' @importFrom cli qty
 #' @importFrom utils packageVersion
+## usethis namespace: end
 NULL
 
 #' @export